tests/wpt/web-platform-tests/html/semantics/forms/autofocus/first-when-later.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>The first autofocus in the document wins, even if elements are inserted later</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#autofocusing-a-form-control:-the-autofocus-attribute">
<link rel="author" title="Domenic Denicola" href="d@domenic.me">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<input autofocus>

<script>
"use strict";

const input1 = document.querySelector("input");

const input2 = document.createElement("input");
input2.autofocus = true;
document.body.appendChild(input2);

step_timeout(() => {
    assert_equals(document.activeElement, input1);
    assert_not_equals(document.activeElement, input2);

    done();
}, 100);
</script>
