LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/autofocus/first-when-later-expected.txt



FAIL The first autofocus in the document wins, even if elements are inserted later. assert_equals: expected Element node <input autofocus=""></input> but got Element node <input autofocus=""></input>
